+++ sleepecg/__init__.py
@@ -10,4 +10,4 @@ __all__ = [
     'rri_similarity',
 ]
 
-__version__ = '0.2.0'
+__version__ = '0.3.0-dev'