+++ sleepecg/io/__init__.py
@@ -1,11 +1,12 @@
 """Functions for downloading and reading datasets."""
 
-from .ecg_readers import read_gudb, read_mitbih
+from .ecg_readers import read_gudb, read_ltdb, read_mitdb
 from .nsrr import download_nsrr, set_nsrr_token
 
 __all__ = [
     'read_gudb',
-    'read_mitbih',
+    'read_ltdb',
+    'read_mitdb',
     'set_nsrr_token',
     'download_nsrr',
 ]