+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## [UNRELEASED]
+
 ## [0.2.0] - 2021-08-11
 ### Added
 - Add interface to download NSRR data ([#8](https://github.com/cbrnr/sleepecg/pull/8) by [Florian Hofer](https://github.com/hofaflo))